move validation logic to a dedicated util file
