+++ src/utils/index.ts
@@ -2,7 +2,11 @@ export { showErrorAlert } from './alertUtils';
 export { formatDate } from './dateUtils';
 export { showErrorModals } from './errors';
 export { Logger } from './Logger';
-export { emailValidation, lengthValidation } from './validation';
+export {
+  emailValidation,
+  lengthValidation,
+  validateProfileForm,
+} from './validation';
 export { getInitials } from './stringUtils';
 export * from './storageUtils';
 export * from './eventEmitter';
